T-9383 program: Check table exists before dropping it
Reviewed-on: https://review.totaralms.com/1180
Tested-by: Jenkins User <jenkins@totaralms.com>

+++ local/db/upgrade.php
@@ -2847,7 +2847,9 @@ function xmldb_local_upgrade($oldversion) {
             if ($result) {
                 // remove management plugin table
                 $table = new XMLDBTable('manager');
-                $result = $result && drop_table($table);
+                if (table_exists($table)) {
+                    $result = $result && drop_table($table);
+                }
 
                 // unset config variables
                 set_config('local_management_version', null);
